Move /beta/percolate tests to new file
